+ Added functions to the BEE class to draw points, lines, and rectangles + Added functions to the BEE class to set and get the drawing color 	! The drawing color is used for the previously mentioned draw functions but is also used to clear the screen + Added get_fontname() to the Font class which retrieves the family and style name as set by the ttf file + Added setters and getters to the InstanceData class which allow the instance's path to be drawn + Added range-checking to the path movement in Room::step_mid() + Made it so that xprevious and yprevious for each instance are actually set in Room::step_mid() * For the InstanceData class: changed xprevious, yprevious, xstart, and ystart to doubles 	* Moved xprevious, yprevious, path_xstart, and path_ystart to public 	* Changed path_current_node from an unsigned int to a regular int * Fixed the get_hspeed() and get_vspeed() functions for the InstanceData class so that they return the speed, not the coordinates * Fixed path_update_node() and handle_path_end() of the InstanceData class so that they actually work * Changed the typedef path_coord in bee/resource_structures/path.hpp from std::tuple<int, int, int> to std::tuple<int, int, double> * Fixed Room::check_paths() to work with reverse speed

+++ bee/resource_structures/path.hpp
@@ -13,7 +13,7 @@
 #include <vector>
 #include <tuple>
 
-typedef std::tuple<int, int, int> path_coord;
+typedef std::tuple<int, int, double> path_coord;
 
 class BEE::Path: public Resource {
 		// Add new variables to the print() debugging method
@@ -42,7 +42,7 @@ class BEE::Path: public Resource {
 		int set_name(std::string);
 		int set_path(std::string);
 		int add_coordinate(path_coord);
-		int add_coordinate(int, int, int);
+		int add_coordinate(int, int, double);
 		int remove_last_coordinate();
 		int remove_coordinate(unsigned int);
 		int set_connection_type(int);
@@ -168,7 +168,7 @@ int BEE::Path::add_coordinate(path_coord new_coordinate) {
 	coordinate_list.push_back(new_coordinate);
 	return 0;
 }
-int BEE::Path::add_coordinate(int x, int y, int speed) {
+int BEE::Path::add_coordinate(int x, int y, double speed) {
 	path_coord new_coordinate (x, y, speed);
 	coordinate_list.push_back(new_coordinate);
 	return 0;
